LayoutTests/imported/w3c/web-platform-tests/scroll-animations/animation-trigger/trigger-scope-scoped-tree-order.tentative-expected.txt

Target
SOURCE 1
SOURCE 2
SOURCE 3
SOURCE 4
SOURCE 5

FAIL Among in-scope triggers with same name, last in tree-order is selected. assert_equals: animation on target has playState paused. expected "paused" but got "running"
